+++ resources/lang/en/admin/overview.php
@@ -1,0 +1,21 @@
+<?php
+
+return [
+    'index' => [
+        'alliance' => 'Alliance',
+        'check_version' => 'Version available',
+        'here' => 'here',
+        'online' => 'Online',
+        'id' => 'ID',
+        'ip' => 'IP address',
+        'last_activity' => 'Last activity',
+        'online_count' => 'Players online',
+        'pm_alt' => '[PM]',
+        'pm_title' => 'Send a PM',
+        'points' => 'Points',
+        'status' => 'Status',
+        'title' => 'Overview',
+        'username' => 'Username',
+        'your_version'  => 'Your version',
+    ],
+];