+++ resources/views/partials/menu.blade.php
@@ -43,7 +43,7 @@
         <tr><td><a href="{{ route('contact') }}" accesskey="3" >@lang('menu.contact')</a></td></tr>
         <tr><td><a href="options.php" accesskey="o">@lang('menu.options')</a></td></tr>
         @if ($user->authlevel > \App\Models\User::LEVEL_PLAYER)
-            <tr><td><a href="admin/overview.php" style="color:lime">{{ trans('user.levels.' . $user->authlevel) }}</a></td></tr>
+            <tr><td><a href="{{ route('admin.index') }}" style="color:lime">{{ trans('user.levels.' . $user->authlevel) }}</a></td></tr>
         @endif
 
         <tr><td><a href="{{ route('user.logout') }}" accesskey="s" style="color:red">@lang('menu.logout')</a></td></tr>